pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.5.2",
    "fastapi[standard]>=0.115.12",
    "grpc-interceptor>=0.15.4",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jose>=1.0.0",
    "passlib>=1.7.4",
    "protobuf-to-pydantic[mypy-protobuf]>=0.3.3.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
